+++ resolver/resolver.go
@@ -2,6 +2,7 @@ package resolver
 
 import (
 	"github.com/docker/docker/api/types"
+	"github.com/docker/docker/api/types/registry"
 	"github.com/docker/docker/client"
 	"golang.org/x/net/context"
 )
@@ -11,6 +12,8 @@ var (
 )
 
 type (
+
+	// Resolver is the main type resolver
 	Resolver struct{}
 
 	systemResolver struct {
@@ -26,21 +29,24 @@ type (
 	systemInfoPluginResolver struct {
 		plugins types.PluginsInfo
 	}
-	// systemInfoRegistryResolver struct {
-	// 	registryConfig registry.ServiceConfig
-	// }
+	systemInfoRegistryResolver struct {
+		registryConfig registry.ServiceConfig
+	}
 )
 
+// System retrieves a docker client
 func (r *Resolver) System() *systemResolver {
 	dkr, _ := client.NewEnvClient()
 	return &systemResolver{system: *dkr}
 }
 
+// Version retrieves server version related info
 func (r *systemResolver) Version() *systemVersionResolver {
 	v, _ := r.system.ServerVersion(ctx)
 	return &systemVersionResolver{version: v}
 }
 
+// Info retrieves a docker host info
 func (r *systemResolver) Info() *systemInfoResolver {
 	i, _ := r.system.Info(ctx)
 	return &systemInfoResolver{info: i}